Refactor BTCChain.transfer in BtcChain to return transaction fee
The `transfer` function in `BtcChain` has been refactored to return the
transaction fee along with the transaction ID. Previously, the function
only returned the transaction ID. This change provides additional
information to the caller about the cost of the transaction. It also
allows for better tracking and analysis of transaction fees. No issues
are affected by this change.

Note: The `tx_vsize_estimate` function has been moved to `utils.rs`.