feat: Update print statement to display signed transaction with fee
Previously, the print statement only displayed the signed transaction in
hexadecimal format. This commit updates the print statement in the
BtcChain struct to include the fee amount in satoshis. By providing this
additional information, it enables easier tracking of the fee associated
with the signed transaction. This change improves the visibility and
transparency of the transaction process. No issues are associated with
this change.

Note: This assumes the `vec_to_hex_string` function is a part of the
codebase and is readily understood by the development team.